withFocusReturn: Convert to TypeScript
