+++ packages/components/tsconfig.json
@@ -50,7 +50,6 @@
 		"src/duotone-picker",
 		"src/gradient-picker",
 		"src/higher-order/with-filters",
-		"src/higher-order/with-focus-return",
 		"src/higher-order/with-notices",
 		"src/navigation",
 		"src/palette-edit"
